Draft of general config-reading facility.
This may be pulled out into fat_core or another gem since it could be used in
almost any command-line app.  It is designed to look for YAML-flavored config
files in the standard place (either the classic places, the XDG places, or
both) and merge them by priority into a Hash object.