Fix problem with recursive --label options.
When one label definition in the labeldb included another label name as one of
its config setting, some options were getting set to nil, notably the
landscape option.

I tracked this down to my use of the :into keyword argument to
OptionParser#parse.  I decided that it was not needed since the @options
instance variable to ArgParser was always where the settings were recorded,
there was no need to pass it in through the :into argument.

For whatever reason, that fixed the problem.